+++ app/javascript/app.vue
@@ -1,25 +1,12 @@
 <template>
-  <v-app id="app">
-    <v-btn>Vuetify</v-btn>
-    <div>
-      <p>{{ message }}</p>
-    </div>
-  </v-app>
+  <div>
+    <router-view />
+  </div>
 </template>
 
 <script>
-export default {
-  data: function () {
-    return {
-      message: "Hello Vue!"
-    }
-  }
-}
 </script>
 
 <style scoped>
-p {
-  font-size: 2em;
-  text-align: center;
-}
 </style>
+
