+++ app/javascript/packs/hello_vue.js
@@ -9,6 +9,7 @@ import Vue from 'vue'
 import Vuetify from "vuetify"
 import "vuetify/dist/vuetify.min.css"
 import App from '../app.vue'
+import router from '../router'
 
 Vue.use(Vuetify)
 const vuetify = new Vuetify();
@@ -16,6 +17,7 @@ const vuetify = new Vuetify();
 document.addEventListener('DOMContentLoaded', () => {
   const app = new Vue({
     vuetify,
+    router,
     render: h => h(App)
   }).$mount()
   document.body.appendChild(app.$el)
