+++ package.json
@@ -8,6 +8,7 @@
     "@rails/webpacker": "5.4.3",
     "vue": "^2.6.11",
     "vue-loader": "^15.8.3",
+    "vue-router": "^3.1.5",
     "vue-template-compiler": "^2.6.14",
     "vuetify": "^2.6.6",
     "webpack": "^4.46.0",
